pyproject.toml
[project]
name = "platanus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "email-validator>=2.0.0",
    "fastapi>=0.121.3",
    "google-generativeai>=0.8.0",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Black configuration
[tool.black]
line-length = 100
target-version = ["py313"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
  | alembic
)/
'''

# Mypy configuration
[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = false
disallow_subclassing_any = false
disallow_untyped_calls = false
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
pretty = true
show_error_codes = true
show_error_context = true

[[tool.mypy.overrides]]
module = [
    "passlib.*",
    "jose.*",
]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--strict-config",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "httpx>=0.28.1",
    "mypy>=1.18.2",
    "pre-commit>=4.4.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.6",
]
